test: cover PottsScorer's reverse orientation, and widen score_thresh to double
The reverse-orientation half of PottsScorer had no suite coverage: clamping
strand under bidirect (the previous commit) left bidirect = TRUE comparing
strand 1 against strand 1, the only bidirect = FALSE test used strand = 1, and
C_gseq_potts scores a string, so it has no fetch, no reverse-complemented
target and no position mapping. Neither the model/rc swap nor the reverse ->
forward coordinate mapping was reachable by anything that ran.

The new test scores bidirect = FALSE, strand = -1 over chr1:2000-2200 for all
four funcs against the oracle's reverse twin, taking the position as the score
at the reported anchor since ties leave the index undefined. Confirmed to fail
six times when the model/rc swap is replaced by an unconditional pair, and once
- the score-at-the-reported-position assertion - when the reverse -> forward
coordinate mapping is disabled. Every other test in the file passes under both
sabotages, which is what makes the point: they all run at strand != -1.

score_thresh is now double end to end. The tables, the accumulator and the
comparison were always double, and the value comes straight from the user, so
rounding it to ~7 digits on the way in was a precision loss with nothing behind
it: with a threshold 1e-7 above the best anchor score, the float form counted
that anchor as being above it. No existing test used a threshold that precise,
so nothing in the suite moves.

The two defensive branches in the anchor clamp are labelled unreachable, with
the reason and with what the i_min = 0 fallback would do if it ever went live -
score anchors outside the interval.

+++ src/PottsScorer.h
@@ -27,13 +27,17 @@ public:
         MOTIF_COUNT         // anchors at or above score_thresh
     };
 
+    // score_thresh is double, unlike PWMScorer's float: the tables, the
+    // accumulator and the comparison in score_interval() are all double, and
+    // this value comes straight from the user, so there is no reason to round
+    // it to ~7 digits on the way in.
     PottsScorer(const PottsModel &model, const std::string &genome_root,
                 bool extend, ScoringMode mode, bool bidirect, char strand,
-                float score_thresh);
+                double score_thresh);
 
     PottsScorer(const PottsModel &model, GenomeSeqFetch *shared_seqfetch,
                 bool extend, ScoringMode mode, bool bidirect, char strand,
-                float score_thresh);
+                double score_thresh);
 
     float score_interval(const GInterval &interval, const GenomeChromKey &chromkey) override;
 
@@ -64,7 +68,7 @@ private:
     PottsModel m_rc;
     ScoringMode m_mode;
     bool m_bidirect;
-    float m_score_thresh;
+    double m_score_thresh;
     double m_last_max_score = -std::numeric_limits<double>::infinity();
 
     // Scratch, reused across calls so a per-bp iterator does not reallocate.
